requirements.txt
python-telegram-bot==20.7
python-docx
pytz
reportlab
python-dotenv
matplotlib